8.1 vrode

+++ src/shared/modules/offer/offer-controller.ts
@@ -1,6 +1,6 @@
 import { inject, injectable } from 'inversify';
 import { Logger } from '../../libs/logger/index.js';
-import { BaseController } from '../../libs/rest/index.js';
+import {AuthorizeMiddleware, BaseController} from '../../libs/rest/index.js';
 import { City, Component } from '../../types/index.js';
 import { Request, Response } from 'express';
 import { OfferService } from './offer-service.interface.js';
@@ -14,25 +14,28 @@ import { ObjectIdValidatorMiddleware } from '../../libs/rest/middleware/object-i
 import { SchemaValidatorMiddleware } from '../../libs/rest/middleware/schema-validator.middleware.js';
 import { createOfferDtoSchema } from './dto-schemas/create-offer-dto.schema.js';
 import { putOfferDtoSchema } from './dto-schemas/put-offer-dto.schems.js';
+import { Config } from '../../libs/config/config.interface.js';
+import { RestSchema } from '../../libs/config/rest.schema.js';
 
 @injectable()
 export class OfferController extends BaseController {
   constructor(
     @inject(Component.Logger) logger: Logger,
     @inject(Component.OfferService) private offerService: OfferService,
+    @inject(Component.Config) private readonly config: Config<RestSchema>,
   ) {
     super(logger);
-    this.addRoute({path: '/premium/:city', method: HttpMethod.Delete, handler: this.indexPremiumForCity.bind(this)});
+    this.addRoute({path: '/premium/:city', method: HttpMethod.Get, handler: this.indexPremiumForCity.bind(this)});
 
-    this.addRoute({path: '/favourite', method: HttpMethod.Get, handler: this.indexFavouriteForUser.bind(this)});
-    this.addRoute({path: '/favourite/:id', method: HttpMethod.Post, handler: this.addToFavourite.bind(this), middlewares: [new ObjectIdValidatorMiddleware(this.offerService, 'id')]});
-    this.addRoute({path: '/favourite/:id', method: HttpMethod.Delete, handler: this.removeFromFavourite.bind(this), middlewares: [new ObjectIdValidatorMiddleware(this.offerService, 'id')]});
+    this.addRoute({path: '/favourite', method: HttpMethod.Get, handler: this.indexFavouriteForUser.bind(this), middlewares: [new AuthorizeMiddleware(this.config.get('JWT_SECRET'))]});
+    this.addRoute({path: '/favourite/:id', method: HttpMethod.Post, handler: this.addToFavourite.bind(this), middlewares: [new ObjectIdValidatorMiddleware(this.offerService, 'id'), new AuthorizeMiddleware(this.config.get('JWT_SECRET'))]});
+    this.addRoute({path: '/favourite/:id', method: HttpMethod.Delete, handler: this.removeFromFavourite.bind(this), middlewares: [new ObjectIdValidatorMiddleware(this.offerService, 'id'), new AuthorizeMiddleware(this.config.get('JWT_SECRET'))]});
 
     this.addRoute({path: '/', method: HttpMethod.Get, handler: this.index.bind(this)});
-    this.addRoute({path: '/', method: HttpMethod.Post, handler: this.create.bind(this), middlewares: [new SchemaValidatorMiddleware(createOfferDtoSchema)]});
+    this.addRoute({path: '/', method: HttpMethod.Post, handler: this.create.bind(this), middlewares: [new SchemaValidatorMiddleware(createOfferDtoSchema), new AuthorizeMiddleware(this.config.get('JWT_SECRET'))]});
     this.addRoute({path: '/:id', method: HttpMethod.Get, handler: this.showById.bind(this), middlewares: [new ObjectIdValidatorMiddleware(this.offerService, 'id')]});
-    this.addRoute({path: '/:id', method: HttpMethod.Put, handler: this.updateById.bind(this), middlewares: [new ObjectIdValidatorMiddleware(this.offerService, 'id'), new SchemaValidatorMiddleware(putOfferDtoSchema)]});
-    this.addRoute({path: '/:id', method: HttpMethod.Delete, handler: this.deleteById.bind(this), middlewares: [new ObjectIdValidatorMiddleware(this.offerService, 'id')]});
+    this.addRoute({path: '/:id', method: HttpMethod.Put, handler: this.updateById.bind(this), middlewares: [new SchemaValidatorMiddleware(putOfferDtoSchema), new ObjectIdValidatorMiddleware(this.offerService, 'id'), new AuthorizeMiddleware(this.config.get('JWT_SECRET'))]});
+    this.addRoute({path: '/:id', method: HttpMethod.Delete, handler: this.deleteById.bind(this), middlewares: [new ObjectIdValidatorMiddleware(this.offerService, 'id'), new AuthorizeMiddleware(this.config.get('JWT_SECRET'))]});
   }
 
   private async index(req: Request, res: Response): Promise<void> {
@@ -57,8 +60,9 @@ export class OfferController extends BaseController {
   }
 
   private async create(req: Request, res: Response): Promise<void> {
+    const { userId } = res.locals;
     const dto = plainToClass(CreateOfferDto, req.body);
-    dto.authorId = new Types.ObjectId();
+    dto.authorId = userId;
     const offer = await this.offerService.create(dto);
     this.created(res, offer);
   }
@@ -81,6 +85,14 @@ export class OfferController extends BaseController {
       this.sendBadRequest('id', id);
     }
 
+    const { userId } = res.locals;
+    const offerId = new Types.ObjectId(id);
+
+    const offerFromDb = await this.offerService.findById(offerId);
+    if (offerFromDb?.authorId !== userId) {
+      throw new HttpError(StatusCodes.FORBIDDEN, 'No access to delete offer');
+    }
+
     const dto = plainToClass(PutOfferDto, req.body);
     dto.id = new Types.ObjectId(id);
     const offer = await this.offerService.change(dto);
@@ -94,7 +106,15 @@ export class OfferController extends BaseController {
       this.sendBadRequest('id', id);
     }
 
-    await this.offerService.deleteById(new Types.ObjectId(id));
+    const { userId } = res.locals;
+    const offerId = new Types.ObjectId(id);
+
+    const offer = await this.offerService.findById(offerId);
+    if (offer?.authorId !== userId) {
+      throw new HttpError(StatusCodes.FORBIDDEN, 'No access to delete offer');
+    }
+
+    await this.offerService.deleteById(offerId);
     this.noContent(res, id);
   }
 
@@ -126,16 +146,51 @@ export class OfferController extends BaseController {
     this.ok(res, offers);
   }
 
-  private async indexFavouriteForUser(_req: Request, _res: Response): Promise<void> {
-    throw new Error('Method not implemented.');
+  private async indexFavouriteForUser(req: Request, res: Response): Promise<void> {
+    const { limit, skip } = req.query;
+
+    const defaultLimit = 20;
+    const limitValue = limit ? parseInt(limit as string, 10) : defaultLimit;
+
+    if (isNaN(limitValue)) {
+      this.sendBadRequest('limit', limit);
+    }
+
+    const defaultSkip = 0;
+    const skipValue = skip ? parseInt(skip as string, 10) : defaultSkip;
+
+    if (isNaN(skipValue)) {
+      this.sendBadRequest('skip', skip);
+    }
+
+    const { userId } = res.locals;
+
+    const offers = await this.offerService.findAllFavourite(userId, limitValue, skipValue);
+    this.ok(res, offers);
   }
 
-  private async addToFavourite(_req: Request, _res: Response): Promise<void> {
-    throw new Error('Method not implemented.');
+  private async addToFavourite(req: Request, res: Response): Promise<void> {
+    const { userId } = res.locals;
+    const { id } = req.params;
+
+    if (!isValidObjectId(id)) {
+      this.sendBadRequest('id', id);
+    }
+
+    await this.offerService.addToFavourite(new Types.ObjectId(id), userId);
+    this.ok(res, null);
   }
 
-  private async removeFromFavourite(_req: Request, _res: Response): Promise<void> {
-    throw new Error('Method not implemented.');
+  private async removeFromFavourite(req: Request, res: Response): Promise<void> {
+    const { userId } = res.locals;
+    const { id } = req.params;
+
+    if (!isValidObjectId(id)) {
+      this.sendBadRequest('id', id);
+    }
+
+    await this.offerService.removeFromFavourite(new Types.ObjectId(id), userId);
+    this.ok(res, null);
   }
 
   private sendBadRequest<T>(paramName: string, value: T): void {
